Permit interrupting compilation via Ctrl+C
In order to avoid that corrupt data (e.g. half of a JSON file) is
written to disk when terminating clcache via a SIGTERM or SIGNINT
handler, the script ignored those signals.

However, this means that calls to clcache can no longer be terminated
via Ctrl+C. In particular, if a cache miss is encountered and a
long-running call to invokeRealCompiler() is performed, there's no way
to abort this.

Hence, this patch adjusts the code such that the signals are only
blocked during cache-modifying phases, namely:

  * a cache miss in direct mode, after the real compiler finished
  * a cache miss in indirect mode, after the real compiler finished
  * cleaning the cache

By just blocking signals for these (comparatively short) phases, we
ensure that no broken object files, no incomplete stdout output and no
malformed JSON files are written to disk when pressing Ctrl+C.
Furthermore, by blocking the signals on this 'high' level, we ensure
that we don't get interrupted between populating the cache and updating
the statistics.

Of course, terminating the process by force still may leave the cache in
an undefined state so more work is needed to make clcache recover
gracefully.